Work around mptt slowness when there are thousands of snippets.
See ticket #85 for an example test case.

This works around mptt slowness via providing the proper defaults
settings, and using the mptt manager to track down the next tree_id
to use.  Note that it's inherintly racey in finding the next tree_id-
this isn't a flaw of this code, it's a flaw of mptt.  If you trace
the _get_next_tree_id() functionality, you'll find that it assumes
that between when that value is computed, and when the insertion happens,
nothing else slipped in- this isn't a safe assumption, although the
timing required to trigger the issue would be fairly tight.